backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Dan <daniel@fail.ie>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.110.0"}
aiohttp = {extras = ["speedups"], version = "^3.9.3"}
tortoise-orm = "^0.20.0"
aerich = "^0.7.2"

[tool.poetry.group.test.dependencies]
pytest-aiohttp = "^1.0.5"
pytest = "^8.0.2"
pytest-mock = "^3.12.0"
aioresponses = "^0.7.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
